+++ rosariosis_mysql.sql
@@ -694,6 +694,7 @@ CREATE TABLE billing_fees (
     waived_fee_id integer,
     created_at timestamp DEFAULT current_timestamp,
     updated_at timestamp NULL ON UPDATE current_timestamp,
+    created_by text,
     FOREIGN KEY (school_id,syear) REFERENCES schools(id,syear)
 );
 
@@ -716,6 +717,7 @@ CREATE TABLE billing_payments (
     file_attached text,
     created_at timestamp DEFAULT current_timestamp,
     updated_at timestamp NULL ON UPDATE current_timestamp,
+    created_by text,
     FOREIGN KEY (school_id,syear) REFERENCES schools(id,syear)
 );